requirements.txt

Flask==2.2.5
Werkzeug==2.2.2
mysql-connector-python==8.0.30
    